refactor: Prefix all iOS classes and structs with `Phone`
This is preparatory work for merging the macOS and iOS targets.
