+++ core/Sources/FileawayCore/Views/Common/ErrorText.swift
@@ -20,29 +20,17 @@
 
 import SwiftUI
 
-import Interact
+public struct ErrorText: View {
 
-import FileawayCore
+    var text: String?
 
-struct DocumentPreviewHeader: View {
-
-    @Binding private var isVisible: Bool
-    private let url: URL
-
-    init(_ isVisible: Binding<Bool>, url: URL) {
-        _isVisible = isVisible
-        self.url = url
+    public init(text: String? = nil) {
+        self.text = text
     }
 
-    var body: some View {
-        ConditionalHeader($isVisible) {
-            DocumentPreviewButton(url: url, size: .init(width: 240, height: 240))
-                .horizontalSpace(.both)
-            Text(url.displayName)
-                .font(.body)
-                .fontWeight(.bold)
-                .multilineTextAlignment(.center)
-        }
+    public var body: some View {
+        Text(text ?? "").foregroundColor(.red)
     }
 
 }
+
