+++ ios/Fileaway/Views/Wizard/PhoneDocumentPreviewButton.swift
@@ -20,25 +20,29 @@
 
 import SwiftUI
 
+import Interact
+
 import FileawayCore
 
-struct ComponentItem: View {
+public struct PhoneDocumentPreviewButton: View {
 
-    @Environment(\.editMode) var editMode
-    @ObservedObject var ruleModel: RuleModel
-    @State var component: ComponentModel
+    @State private var previewURL: URL?
 
-    var body: some View {
-        HStack {
-            if component.type == .text {
-                EditText("Component", text: $component.value).environment(\.editMode, editMode)
-            } else {
-                Text(ruleModel.name(for: component))
-                    .tokenAppearance()
-                    .tint(component.variable?.color ?? .black)
-            }
+    private let url: URL
+    private let size: CGSize
+
+    init(url: URL, size: CGSize) {
+        self.url = url
+        self.size = size
+    }
+
+    public var body: some View {
+        Button() {
+            previewURL = url
+        } label: {
+            IconView(url: url, size: size)
         }
-        .environment(\.editMode, editMode)
+        .quickLookPreview($previewURL)
     }
 
 }
